Debt/620 angular2 charactergen (#2016)
* Make the item description the heading instead of name

Like how heading is character summary

* starting

* starting on character

* working on character formatter

* working on it

* getting close

split the spell grouping into a separate service, getting the formatter tests passing

* working on leadership service tests now

* working on character component tests

Also added a BonusPipe to handle +/- signs for numbers

* working through it

introduced the bonuses pipe to help with formatting things

enhanced the bonus pipe to handle the conditional asterisk as well

* Working on skills table now

Have a measurement pipe now, really simplifies a lot of the formatting stuff

Maybe change the formatter services into pipes? It's the same concept...

* skills in progress

* Starting to move formatting services to pipes

Pipes make more sense and are much cleaner. Doesn't build right now, but switching to other things

* pausing again

* all format services are now pipes

* Update leadership.service.spec.ts

* weird test issue

feat tests are failing, but not sure why. Must need some sort of clarity on the selectors

* starting on spell quantities

* Hugo is up

* working on equipment stuff

Also sort skills in the component

* Update character.component.spec.ts

* Leadership component done

Now just to work on the charactergen component

* stopping

* working on tests

* roll expression examples

* more tests

* Hugo woke up

* Update charactergen.component.ts

* stopping for now

* loading component

* working on charactergen html

* Update charactergen.component.html

* a bit more

* working on it

* Update charactergen.component.spec.ts

* working to get tests passing now

* Starting creating the test helper

* more helper stuff

* still combing through the mess

* Update charactergen.component.spec.ts

* stopping for today

* observable chaining not working in charactergen

* getting the repeats to work is tricky

currently it's repeating the last tap, but not the requests or generating message setting

* unit tests passing

now to debug the integration tests for charactergen

* gotta stop for the day

* only 1 test failing

it is being stubborn

* some more

* All passing!

* Update charactergen.component.spec.ts

* Fixed?

* Update charactergen.component.spec.ts

* Update charactergen.component.spec.ts

* Update charactergen.component.spec.ts

* Did manual testing

* Made some formatting adjustments to code and tests
* Added Safari and Edge to CI tests, for giggles
* Cleaned up README
* Debounce on expression typing

* Update package.json

* remove safari and edge

* Update treasuregen.component.spec.ts

* Update packages

* Having weird pipeline issues

Wonder if it's a parallelism problem. Sequencing some of the jobs to see if it helps

* Reorder jobs

1. Build APIs
2. Build Websites
3. Export Post-Deployment Tests

Undo the retry for static web tests, let the tests be more robust

* Update build.yml

* Remove ordering

The UseDotNet step should fix the issue, ordering not needed, can run all in parallel again. Also remove the UseDotNet from the static website, as it doesn't need it

Also refine some more flaky tests

* Update charactergen.component.spec.ts

* Extend karma browser timeout

Extend to 3.5 minutes (was defaulted at 1 minute)

Also clean up some spec names

* Update karma.conf.ci.js

* Update karma.conf.ci.js

* Update karma.conf.ci.js

3 minutes hit timeouts, try 4 minutes

* Fix null warning